fix(stats): DiscreteDistribution on the signed integers with support(), as SciPy (szq1n.8)
The trait was u64-only: negative k was unreachable, and Skellam / DiscreteLaplace `rvs` cast
signed samples with `as u64`, so −1 became 18446744073709551615 and the mean of Skellam(2, 3)
draws through the trait was ~1e18 instead of −1. Their trait ppf/isf returned NaN.

- pmf / logpmf / cdf / sf / logcdf / logsf take `i64`; `rvs` returns `Vec<i64>`; `fit` / `try_fit`
  take `&[i64]` (a family on k >= 0 refuses a negative observation with UnsupportedData instead
  of wrapping it). New required `support() -> (f64, f64)`, SciPy's `support()` for all 18 impls.
- Non-negative families: a guard answers below the support as SciPy does (pmf 0, cdf 0, sf 1,
  logpmf/logcdf −inf, logsf 0) and the existing body runs unchanged, so every k >= 0 output is
  bit-identical. The default cdf / ppf / isf searches keep the unsigned start for supports at or
  above 0.
- Skellam: SciPy's negative branch cdf(k<0) = chndtr(2μ2, −2k, 2μ1), sf = 1 − that; logpmf for
  either sign (log I_|k|, and the degenerate μ1 = 0 / μ2 = 0 cases on the correct side).
- DiscreteLaplace: pmf/logpmf in |k|, SciPy's e^{a(k+1)}/(e^a+1) for cdf at k < 0.
- ppf / isf on a support unbounded below: bracket outward from the rounded mean and bisect, so
  skellam and dlaplace now have SciPy's quantiles (they returned NaN).
- ppf(1) and isf(0) return the support's upper bound (SciPy rv_discrete), not +inf for a
  bounded family.
- RandInt implements the trait (support [low, high − 1], SciPy's entropy and kurtosis).
- Callers, bins, bench, metamorphic tests and 16 conformance tests follow the signed argument.

Tests: discrete_trait_reaches_negative_support_like_scipy — skellam(2,3) pmf/cdf/sf/logpmf at
k = −5..2, ppf/isf at five quantiles, dlaplace(0.8) likewise (SciPy 1.17.1 values), the negative
case the old trait failed (100 000 seeded draws: mean within 0.02 of −1, negatives present),
signed fit round trip, Poisson below its support and refusing a negative observation, finite
supports; conformance diff_stats_discrete_signed_support.rs (skellam, dlaplace, randint(−3, 3)
via the trait, 126 rows, live SciPy) + tolerance anchor. discrete_ppf_isf_match_scipy asserted
the old NaN ppf/isf for skellam/dlaplace; it now asserts SciPy's values (skellam(2,1).ppf(0.5) = 1,
ppf(0.05) = −2, dlaplace(0.5).isf(0.5) = 0, isf(0.9) = −3).

Sidecar: tolerance_baseline.json changed (the new anchor, and 6a9f36c4c's anchors before it, which
left its RaptorQ sidecar stale — CI G0 and G8 both failed on that source_hash). Regenerated with
the repo's raptorq_sidecar built on vmi1227854 (binary sha256 a389202e…16db, copied back by scp);
`raptorq_sidecar --verify` said FAIL (source_hash mismatch) before and ok after. The sidecar is a
durability record of the source bytes, not a behavioural golden.

Evidence (remote rch, vmi1227854): fsci-stats lib 2199 passed + metamorphic 60 passed (b15);
cargo check of every fsci-stats target and the 20 conformance targets that touch discrete families
exit 0 (b13); pinned clippy -D warnings exit 0 for fsci-stats + fsci-linalg --all-targets and for
the fsci-conformance lib + edited test targets (b15). The live-SciPy diff is CI G3.

+++ crates/fsci-conformance/tests/diff_stats_bernoulli.rs
@@ -207,7 +207,7 @@ fn diff_stats_bernoulli() {
         let oracle = pmap.get(&case.case_id).expect("validated oracle");
         let dist = Bernoulli::new(case.p);
         if let Some(spmf) = oracle.pmf {
-            let d = (dist.pmf(case.k) - spmf).abs();
+            let d = (dist.pmf(case.k as i64) - spmf).abs();
             max_overall = max_overall.max(d);
             diffs.push(CaseDiff {
                 case_id: case.case_id.clone(),
@@ -217,7 +217,7 @@ fn diff_stats_bernoulli() {
             });
         }
         if let Some(scdf) = oracle.cdf {
-            let d = (dist.cdf(case.k) - scdf).abs();
+            let d = (dist.cdf(case.k as i64) - scdf).abs();
             max_overall = max_overall.max(d);
             diffs.push(CaseDiff {
                 case_id: case.case_id.clone(),
